Fix historical legacy issues and command registration problems. Config. yml file name issue
